views: Add support for boolean 'series' parameters
Previously, we allowed users to download patch mboxes with dependencies
included using a 'series' parameter. This accepted either a numeric ID,
corresponding to the ID of the patch series that dependencies should be
included from, or a wildcard value ('*').

    /patch/{patchID}/mbox/?series=123
    /patch/{patchID}/mbox/?series=*

With switch to a 1:N series-patch relationship, this is clearly
unnecessary now but must be retained to avoid breaking users. However,
that doesn't mean we can't things a little clearer. Add support for
boolean parameters, which make more sense for this kind of relationship:

  /patch/{patchID}/mbox/?series=true
  /patch/{patchID}/mbox/?series=1
  /patch/{patchID}/mbox/?series=false
  /patch/{patchID}/mbox/?series=0

Signed-off-by: Stephen Finucane <stephen@that.guru>