Rework tagging infrastructure
Solve #113 and #57 GitHub issues, keep track of tag origin to be able
to add tags to comments in the API later.

Tags are now related to series rather then patches. If the tag comes
from a cover letter, only the series field is set, if from a comment on
cover, both series and comment fields are set, if from a patch, series
and patch fields are set and if from a comment on a patch, all fields
are set.

Move functionality related to tag extraction into parser.py.

Signed-off-by: Veronika Kabatova <vkabatov@redhat.com>