Add help text to project and patch_project fields
Duplication of the field to avoid performance issues caused some
unexpected results when moving patches between projects in the admin
interface. It's easy to change the "project" field and click save,
instead of double checking which field needs to be modified and kept in
sync with the one being changed. This lead to patch showing correct
project in the API and patch detail in the web UI, but the patch itself
was listed in the wrong project when looking at the patch listings.

Because of the discussions of the denormalization of the tables, trying
to code automatic modification of the other field if one is being
changed seems like too much work for very little benefit. What we can do
instead is mention this dependency in fields' help text. Modification of
the project is not an everyday task so it shouldn't put too much burden
on the users and this simple reminder can prevent unexpected breakage
and bug reports.

Signed-off-by: Veronika Kabatova <vkabatov@redhat.com>