Only mark series as completed after cover letter has arrived
People often put metadata into the cover letter. These are needed by
various CI systems built upon Patchwork, and they are unable to process
the series if a cover letter is expected to arrive and it didn't arrive
at the time of processing. Email delaying is normal and the CI systems
are currently unable to exclude series based on the `received_all`
attribute.

Add a new DB field (not exposed in the API) that says "a cover letter
should arrive but it didn't yet" and use this field when assessing
whether the series are completed or not. In the end, a cover letter is
a vital part of the series.

Signed-off-by: Veronika Kabatova <vkabatov@redhat.com>